requirements.txt
PyYAML==5.3.1
allure-pytest==2.9.45
allure-python-commons==2.9.45
attrs==21.4.0
pandas==1.4.2
pytest==7.0.0
requests==2.25.1
pymysql==0.9.3